+++ test/ffmpeg-vaapi/encode/jpeg.py
@@ -6,28 +6,10 @@
 
 from ....lib import *
 from ....lib.ffmpeg.vaapi.util import *
-from ....lib.ffmpeg.vaapi.encoder import EncoderTest
-from ....lib.codecs import Codec
+from ....lib.ffmpeg.vaapi.encoder import JPEGEncoderTest
 
 spec      = load_test_spec("jpeg", "encode")
 spec_r2r  = load_test_spec("jpeg", "encode", "r2r")
-
-@slash.requires(*have_ffmpeg_encoder("mjpeg_vaapi"))
-@slash.requires(*platform.have_caps("vdenc", "jpeg"))
-class JPEGEncoderTest(EncoderTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      caps  = platform.get_caps("vdenc", "jpeg"),
-      codec = Codec.JPEG,
-      ffenc = "mjpeg_vaapi",
-    )
-
-  def get_file_ext(self):
-    return "mjpeg" if self.frames > 1 else "jpg"
-
-  def get_vaapi_profile(self):
-    return "VAProfileJPEGBaseline"
 
 class cqp(JPEGEncoderTest):
   def init(self, tspec, case, quality):
